database geseed

+++ database/migrations/2024_04_08_092951_create_reservering_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('baanId');
             $table->integer('pakketOptieId')->nullable();
             $table->integer('reserveringStatusId');
-            $table->integer('reserveringsNummer');
+            $table->bigInteger('reserveringsNummer');
             $table->date('datum');
             $table->integer('aantalUren');
             $table->time('beginTijd');
